fix(store,command,persist): stream correctness and RDB type tag
Stream fixes:
- XCLAIM/XAUTOCLAIM: use < instead of <= for min-idle-time (match Redis)
- XADD: track high watermark in stream_last_ids to maintain monotonic IDs
  after XDEL removes all entries
- XREADGROUP: pending replay no longer increments delivery count (match Redis)
- AOF: only emit XSETID when watermark exceeds max entry ID

RDB:
- Rename ZSET type constant to ZSET_2 to document binary LE double encoding

+++ crates/fr-command/src/lib.rs
@@ -17329,17 +17329,18 @@ mod tests {
         assert_eq!(
             detail,
             RespFrame::Array(Some(vec![
+                // Pending replay does NOT increment delivery count or update idle time
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-0".to_vec())),
                     RespFrame::BulkString(Some(b"c1".to_vec())),
-                    RespFrame::Integer(20),
-                    RespFrame::Integer(2),
+                    RespFrame::Integer(30), // idle since initial delivery at t=10
+                    RespFrame::Integer(1),  // delivery count stays at 1
                 ])),
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-1".to_vec())),
                     RespFrame::BulkString(Some(b"c1".to_vec())),
-                    RespFrame::Integer(20),
-                    RespFrame::Integer(2),
+                    RespFrame::Integer(30),
+                    RespFrame::Integer(1),
                 ])),
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-2".to_vec())),
@@ -17370,14 +17371,14 @@ mod tests {
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-0".to_vec())),
                     RespFrame::BulkString(Some(b"c1".to_vec())),
-                    RespFrame::Integer(20),
-                    RespFrame::Integer(2),
+                    RespFrame::Integer(30),
+                    RespFrame::Integer(1),
                 ])),
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-1".to_vec())),
                     RespFrame::BulkString(Some(b"c1".to_vec())),
-                    RespFrame::Integer(20),
-                    RespFrame::Integer(2),
+                    RespFrame::Integer(30),
+                    RespFrame::Integer(1),
                 ])),
             ]))
         );
